backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pyyaml==6.0.1
aiofiles==23.2.1
websockets==12.0
psutil==5.9.6
pydantic==2.5.0
python-dotenv==1.0.0
Pillow>=10.0.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
